Warnings cleanup

+++ dodSpaceInvaders/src/dod/MemPool.cpp
@@ -2,7 +2,7 @@
 
 Dod::MemPool::MemPool(MemTypes::capacity_t capacityInBytes) noexcept
 {
-	this->data.resize(capacityInBytes);
+	this->data.resize(static_cast<size_t>(capacityInBytes));
 	this->dataBegin = this->data.data();
 	this->dataEnd = this->data.data() + this->data.size();
 }
